Remove nullability on Comparator.compare arguments
I think this is possible due to explanations in https://kotlinlang.org/docs/reference/java-interop.html#null-safety-and-platform-types

The arguments are platform types, and to be safe IntelliJ generates them as Int? but it's allowed to denote them as Int, with runtime checks.

Seems like a good pragmatic approach, definitely works here